airflow/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.3
anyio==3.6.2
apache-airflow==2.5.3
apache-airflow-providers-common-sql==1.4.0
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.3.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
appdirs==1.4.4
argcomplete==3.0.6
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.6.2
boto3==1.26.118
botocore==1.29.118
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.14
cryptography==40.0.2
Deprecated==1.2.13
dill==0.3.6
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozendict==2.3.7
frozenlist==1.3.3
graphviz==0.20.1
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httpx==0.24.0
idna==3.4
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.3
packaging==23.1
pandas==2.0.0
pathspec==0.9.0
pendulum==2.1.2
pluggy==1.0.0
prison==0.2.1
psutil==5.9.5
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.6.0
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rich==13.3.4
s3transfer==0.6.0
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.0
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.1
yfinance==0.2.18
